Check for both files, but provide warnings

+++ entrypoint.sh
@@ -34,11 +34,22 @@ git config --global user.name "$INPUT_AUTHOR"
 git clone --single-branch --branch "$INPUT_TARGET_BRANCH" "https://$INPUT_TOKEN@github.com/$INPUT_DESTINATION_REPO.git" "$CLONE_DIR"
 ls -la "$CLONE_DIR"
 
-echo "Copying contents to to git repo"
+echo "Copying contents to to git repo IF THEY EXIST"
 # Include dot files for source filepath
 mkdir -p $CLONE_DIR/$INPUT_DESTINATION_FOLDER
-cp -r "$INPUT_SOURCE_FILE_PATH"/* "$CLONE_DIR/$INPUT_DESTINATION_FOLDER"
-cp -r "$INPUT_SOURCE_FILE_PATH"/.??* "$CLONE_DIR/$INPUT_DESTINATION_FOLDER"
+
+if test -f "$INPUT_SOURCE_FILE_PATH"/*; then
+  cp -r "$INPUT_SOURCE_FILE_PATH"/* "$CLONE_DIR/$INPUT_DESTINATION_FOLDER"
+else
+  echo "WARNING: No visible files exist"
+fi
+invisible_exists=false
+if test -f "$INPUT_SOURCE_FILE_PATH"/.??*; then
+  cp -r "$INPUT_SOURCE_FILE_PATH"/.??* "$CLONE_DIR/$INPUT_DESTINATION_FOLDER"
+else
+  echo "WARNING: No invisible/hidden (dot) files exist"
+fi
+
 cd "$CLONE_DIR"
 ls -la
